device_control_backend/requirements.txt
requests
psycopg2
asyncio
pyzmq
slackclient
pytest
